% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaElement.R
\name{getMetaElement}
\alias{getMetaElement}
\title{Get the metadata of the meteorlogical element for a specific time resolution}
\usage{
getMetaElement(timeserietypeID, stations = NULL)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{stations}{id of the stations. Can be NULL, a scalar or a vector of station ids}
}
\description{
The function \code{getMetaElement()} makes it possible to get the metadata of meteorological elements measured by met.no's stations
}
\details{
The output is a dataframe. This structure eases the conversion into json
}
\examples{
tmp <- getMetaElement(timeserietypeID="2")
str(tmp)
tmp <- getMetaElement(timeserietypeID="0",stations=180)
str(tmp)
}
\keyword{wsklimaR}

