% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetaStationbyElem.R
\name{getMetaStationbyElem}
\alias{getMetaStationbyElem}
\title{Get the metadata of the stations given several meteorlogical element}
\usage{
getMetaStationbyElem(timeserietypeID, elements)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Hourly, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{meteorlogical element as its code ("TAM", "RR", ...). Can be a vector of elements}
}
\description{
The function \code{getMetaStationbyElem()} makes it possible to get the metadata of stations, registered by met.no, that measure all the chosen meteorological elements.
}
\details{
The output is a dataframe. This structure eases the conversion into json
}
\examples{
tmp <- getMetaStationbyElem(timeserietypeID="2",elements="TA")
str(tmp)
tmp <- getMetaStationbyElem(timeserietypeID="0",elements=c("TAM","RR"))
str(tmp)
}
\keyword{wsklimaR}
