% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.checkDateTS.R
\name{wsklimaR.checkDateTS}
\alias{wsklimaR.checkDateTS}
\title{Check the data timeseries and fill-in by missing values}
\usage{
wsklimaR.checkDateTS(timeserie, timeserietypeID, fromPeriod, toPeriod,
  missingValues)
}
\arguments{
\item{timeserie}{a 2-column data.frame with time and the values}

\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{fromPeriod}{string of character yyyymmdd}

\item{toPeriod}{string of character yyyymmdd}

\item{missingValues}{value of the expected mssing values.}
}
\description{
In case of missing dates, the date timeserie will get filled-in, and a missing value will be added as a value of the element
}
\examples{
\dontrun{
tmp <- wsklimaR.checkDateTS(timeserie, timeserietypeID, fromPeriod, toPeriod, missingValues)
}

}
\keyword{wsklimaR}

