% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getDateTS.R
\name{wsklimaR.getDateTS}
\alias{wsklimaR.getDateTS}
\title{Get the full date timeserie}
\usage{
wsklimaR.getDateTS(fromPeriod, toPeriod, timeserietypeID)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}
}
\description{
Get the full date timeserie between the range of the period and at the right time resolution
}
\examples{
wsklimaR.getDateTS(fromPeriod,toPeriod,timeserietypeID)
}
\keyword{wsklimaR}

