% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getMetaElement.R
\name{wsklimaR.getMetaElement}
\alias{wsklimaR.getMetaElement}
\title{Get the metadata of the meteorlogical element for a specific time resolution}
\usage{
wsklimaR.getMetaElement(timeserietypeID, station = NULL)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{TAM, RR, ..}
}
\description{
wsklimaR.getMetaElement makes it possible to get the metadata of meteorological elements measured by met.no's stations
}
\details{
The output is a dataframe. This structure eases the conversion into json
}
\examples{
tmp <- wsklimaR.getMetaElement(timeserietypeID="2")
str(tmp)
tmp <- wsklimaR.getMetaElement(timeserietypeID="0",station=180)
str(tmp)
}
\keyword{wsklimaR}

