% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getMetaSbyE.R
\name{wsklimaR.getMetaSbyE}
\alias{wsklimaR.getMetaSbyE}
\title{Get the metadata of the stations given a meteorlogical element}
\usage{
wsklimaR.getMetaSbyE(timeserietypeID, elements)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{meteorlogical element as its code ("TAM", "RR", ...). Can be a vector if multi-element}
}
\description{
wsklimaR.getMetaSbyE makes it possible to get the metadata of the stations from met.no given a meteorological element.
It is not recommended to use this function to get metadata of several elements.
Indeed, instead of the logic operator AND, wsklima uses the logic operator OR
It is highly recommended to use the function wsklimaR.getMetaStationbyElement().
}
\examples{
tmp <- wsklimaR.getMetaSbyE(timeserietypeID="0",elements="TAM")
}
\keyword{wsklimaR}

