% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getMetaStationbyElement.R
\name{wsklimaR.getMetaStationbyElement}
\alias{wsklimaR.getMetaStationbyElement}
\title{Get the metadata of the stations given several meteorlogical element}
\usage{
wsklimaR.getMetaStationbyElement(timeserietypeID, elements)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{TAM, RR, ..}
}
\description{
wsklimaR.getMetaStationbyElement makes it possible to get the metadata of the met.no's stations that measure all the chosen meteorological elements.
}
\details{
The output is a dataframe. This structure eases the conversion into json
}
\examples{
tmp <- wsklimaR.getMetaStationbyElement(timeserietypeID="2",elements="TA")
str(tmp)
tmp <- wsklimaR.getMetaStationbyElement(timeserietypeID="0",elements=c("TAM","RR"))
str(tmp)
}
\keyword{wsklimaR}

