% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getVByMonthCompleteTS.R
\name{wsklimaR.getVByMonthCompleteTS}
\alias{wsklimaR.getVByMonthCompleteTS}
\title{Get the complete timeseries for one meteorlogical element at one station by month}
\usage{
wsklimaR.getVByMonthCompleteTS(timeserietypeID, fromPeriod, toPeriod, stations,
  elements, missingValues)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{fromPeriod}{string of character yyyymmdd}

\item{toPeriod}{string of character yyyymmdd}

\item{stations}{id of the stations.}

\item{elements}{meteorlogical element as its code ("TAM", "RR", ...). Can be a vector if multi-element.}

\item{missingValues}{value of the expected mssing values.}
}
\description{
The processs is split by month of the period in order to make the wsklima server not to send you an error
The timeserie is checked and filling-in by missing values in case of missing dates
}
\examples{
wsklimaR.getVByMonthCompleteTS(timeserietypeID="1",
                               fromPeriod="20130101",
                               toPeriod="20130105",
                               stations="180",
                               elements="TAM",
                               missingValues=-999)
}
\keyword{wsklimaR}

