% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getVCompleteTS.R
\name{wsklimaR.getVCompleteTS}
\alias{wsklimaR.getVCompleteTS}
\title{Get the complete timeseries fo one element at one station}
\usage{
wsklimaR.getVCompleteTS(timeserietypeID, fromPeriod, toPeriod, stations,
  elements, missingValues)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{TAM, RR, ...}
}
\description{
The timeserie is checked and filling-in by missing values in case of missing dates
}
\examples{
wsklimaR.getVCompleteTS(timeserietypeID="1", fromPeriod="20130101",toPeriod="20130105",stations="180",elements="TAM",missingValues=-999)
}
\keyword{wsklimaR}

