% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getVMultiStationMultiElement.R
\name{wsklimaR.getVMultiStationMultiElement}
\alias{wsklimaR.getVMultiStationMultiElement}
\title{Get the values of several element at several station}
\usage{
wsklimaR.getVMultiStationMultiElement(timeserietypeID, fromPeriod, toPeriod,
  stations, elements, missingValues)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{TAM, RR, ...}
}
\description{
Get the values of several element at several station
}
\examples{
wsklimaR.getVMultiStationMultiElement(timeserietypeID="4", fromPeriod="20130101",toPeriod="20130105",stations="180",elements="TAM",missingValues=-999)
}
\keyword{wsklimaR}

