% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.checkDateTS.R
\name{wsklimaR.checkDateTS}
\alias{wsklimaR.checkDateTS}
\title{Check the data timeseries and fill-in by missing values}
\usage{
wsklimaR.checkDateTS(timeserie, timeserietypeID, fromPeriod, toPeriod,
  missingValues)
}
\arguments{
\item{timeserie}{a 2-column data.frame with time and the values}
}
\description{
In case of missing dates, the date timeserie will get filled-in, and a missing value will be added as a value of the element
}
\examples{
wsklimaR.checkDateTS(timeserie,toPeriod,timeserietypeID)
}
\keyword{wsklimaR}

