% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getV.R
\name{wsklimaR.getV}
\alias{wsklimaR.getV}
\title{Get the values of one elements at one station}
\usage{
wsklimaR.getV(timeserietypeID, fromPeriod, toPeriod, stations, elements)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{fromPeriod}{string of character yyyymmdd}

\item{toPeriod}{string of character yyyymmdd}

\item{stations}{id of the stations.}

\item{elements}{meteorlogical element as its code ("TAM", "RR", ...). Can be a vector if multi-element}
}
\description{
wsklimaR.getV makes it possible to get timeseries of a meteorlogical element at a given station registered by met.no for a specific period of time and a specific time resolution.
In case of missing dates, the date timeserie will be incomplete.
It is not recommended to use this function to get several elements at different stations.
Indeed, some timeseries are incomplete and thus make the gathering of the different timeseries into a regular data.frame not possible.
To check and fill-in missing dates, please use the function wsklimaR.checkDateTimeSerie(). It makes sure the timeserie to be complete
It is highly recommended to use the function wsklimaR.getValues().
}
\examples{
tmp <- wsklimaR.getV(timeserietypeID="1",
                     fromPeriod="20130101",
                     toPeriod="20130105",
                     stations="180",
                     elements="TAM")
}
\keyword{wsklimaR}

