% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wsklimaR.getVByMonth.R
\name{wsklimaR.getVByMonth}
\alias{wsklimaR.getVByMonth}
\title{Get the values of one elements at one station by month}
\usage{
wsklimaR.getVByMonth(timeserietypeID, fromPeriod, toPeriod, stations, elements)
}
\arguments{
\item{timeserietypeID}{0:Daily, 1:Monthly, 2:Observations-hourly based, 3:Normal Monthly, 4: Normal Daily, 5:Record Daily (not used)}

\item{elements}{TAM, RR, ...}
}
\description{
The wsklima server is not very young. It is better to split the request into several shorter in order to keep the server stable
}
\examples{
wsklimaR.getVByMonth(timeserietypeID="1", fromPeriod="20130101",toPeriod="20130105",stations="180",elements="TAM")
}
\keyword{wsklimaR}

